+++ extLibs/uil/src/main/java/com/nostra13/universalimageloader/utils/IoUtils.java
@@ -93,9 +93,7 @@ public final class IoUtils {
         if (listener != null) {
             boolean shouldContinue = listener.onBytesCopied(current, total);
             if (!shouldContinue) {
-                if (100 * current / total < CONTINUE_LOADING_PERCENTAGE) {
-                    return true; // if loaded more than 75% then continue loading anyway
-                }
+                return 100 * current / total < CONTINUE_LOADING_PERCENTAGE; // if loaded more than 75% then continue loading anyway
             }
         }
         return false;