+++ library/src/main/java/candybar/lib/fragments/RequestFragment.java
@@ -402,9 +402,6 @@ public class RequestFragment extends Fragment implements View.OnClickListener {
             while (!isCancelled()) {
                 try {
                     Thread.sleep(2);
-                    /*
-                    Intent intent = new Intent(Intent.ACTION_SENDTO, Uri.fromParts("mailto", getResources().getString(R.string.dev_email), null));
-                    */
 
                     boolean nonMailingAppSend = getResources().getBoolean(R.bool.enable_non_mail_app_request);
                     Intent intent;
@@ -581,7 +578,7 @@ public class RequestFragment extends Fragment implements View.OnClickListener {
 
                 // Checks If Request can be Done or Not
                 if (appVersionCode < disabledReqBelow) canRequest = false;
-                String disabledReqOns[] = disabledReqOn.split("[\\s,]");
+                String[] disabledReqOns = disabledReqOn.split("[\\s,]");
                 for (String version : disabledReqOns) {
                     //Log.d("Disabled Versions", version);
                     //Log.d("Disabled Version Length", version.length() + "");