+++ library/src/main/java/candybar/lib/helpers/DeviceHelper.java
@@ -46,13 +46,10 @@ public class DeviceHelper {
         } catch (PackageManager.NameNotFoundException ignored) {
         }
 
-
-        String installationFlag;
-
         if (installerPackage == null || !installerPackage.contentEquals("com.android.vending")) {
-            installationFlag = "";
+            appVersion = context.getResources().getString(R.string.version_outside_playstore).replace("{{appVersion}}", appVersion);
         } else {
-            installationFlag = " R";
+            appVersion = context.getResources().getString(R.string.version_from_playstore).replace("{{appVersion}}", appVersion);
         }
 
         sb.append("Manufacturer : ").append(Build.MANUFACTURER)
@@ -61,7 +58,7 @@ public class DeviceHelper {
                 .append("\nScreen Resolution : ")
                 .append(width).append(" x ").append(height).append(" pixels")
                 .append("\nAndroid Version : ").append(Build.VERSION.RELEASE)
-                .append("\nApp Version : ").append(appVersion).append(installationFlag)
+                .append("\nApp Version : ").append(appVersion)
                 .append("\nCandyBar Version : ").append(BuildConfig.VERSION_NAME)
                 .append("\n");
         return sb.toString();